Remove unused fields #26
