+++ src/FormBuilder.Domains/Files/Commands/DeleteFile/DeleteFileCommand.cs
@@ -1,20 +1,8 @@
-using FluentValidation;
 using MediatR;
 
 namespace FormBuilder.Domains.Files.Commands.DeleteFile;
 
 public class DeleteFileCommand : IRequest<Unit>
 {
-    public string Uri { get; set; }
-}
-
-public class DeleteFileCommandValidator : AbstractValidator<DeleteFileCommand>
-{
-    public DeleteFileCommandValidator()
-    {
-        RuleFor(x => x.Uri)
-            .NotNull()
-            .NotEmpty()
-            .WithMessage(payload => $"Uri is required");
-    }
+    public string Uri { get; set; } = string.Empty;
 }
