+++ src/FormBuilder.Domains/Files/Commands/DeleteFile/DeleteFileCommandValidator.cs
@@ -1,0 +1,14 @@
+﻿using FluentValidation;
+
+namespace FormBuilder.Domains.Files.Commands.DeleteFile;
+
+public class DeleteFileCommandValidator : AbstractValidator<DeleteFileCommand>
+{
+    public DeleteFileCommandValidator()
+    {
+        RuleFor(x => x.Uri)
+            .NotNull()
+            .NotEmpty()
+            .WithMessage(payload => $"Uri is required");
+    }
+}
