+++ src/FormBuilder.Domains/Files/Queries/GetFileByUri/GetFileByUriQuery.cs
@@ -1,21 +1,9 @@
-using FluentValidation;
 using FormBuilder.Domains.Files.Models;
 using MediatR;
 
 namespace FormBuilder.Domains.Files.Queries.GetFileByUri;
 
-public class GetFileByUriQuery :IRequest<DownloadFileMediaModel>
+public class GetFileByUriQuery : IRequest<DownloadFileMediaModel>
 {
-    public string Uri { get; set; }
-}
-
-public class GetFileByUriQueryValidator : AbstractValidator<GetFileByUriQuery>
-{
-    public GetFileByUriQueryValidator()
-    {
-        RuleFor(x => x.Uri)
-            .NotNull()
-            .NotEmpty()
-            .WithMessage(payload => $"Uri is required");
-    }
+    public string Uri { get; set; } = string.Empty;
 }
