+++ src/FormBuilder.Domains/Forms/Commands/UpdateForm/UpdateFormCommand.cs
@@ -7,10 +7,7 @@ public class UpdateFormCommand : IRequest<FormModel>
 {
     public Guid Id { get; set; }
 
-    public string Title { get; set; }
-
-    // [Obsolete]
-    // public string Content { get; set; }
+    public string Title { get; set; } = string.Empty;
 
     public IList<FormItemModel> Items { get; set; } = new List<FormItemModel>();
