+++ src/FormBuilder.Domains/Forms/Models/FormItemOptionModel.cs
@@ -23,7 +23,7 @@ public class FormItemOptionLocaledModel
 
     public Guid? LanguageId { get; set; }
 
-    public string LanguageCode { get; set; }
+    public string LanguageCode { get; set; } = string.Empty;
 
-    public string Text { get; set; }
+    public string Text { get; set; } = string.Empty;
 }
