+++ src/FormBuilder.Domains/Languages/Models/LanguageModel.cs
@@ -3,10 +3,10 @@ namespace FormBuilder.Domains.Languages.Models;
 public class LanguageModel
 {
     public Guid Id { get; set; }
-    
-    public string Code { get; set; }
-    
-    public  string Name { get; set; }
-    
-    public  int Ordinal { get; set; }
+
+    public string Code { get; set; } = string.Empty;
+
+    public string Name { get; set; } = string.Empty;
+
+    public int Ordinal { get; set; } = 1;
 }
