+++ src/FormBuilder.Domains/Results/Commands/AddResult/AddResultCommandHandler.cs
@@ -23,14 +23,12 @@ public class AddResultCommandHandler : IRequestHandler<AddResultCommand, ResultM
         var result = new Result
         {
             FormId = request.FormId,
-            // TODO: remove
-            Content = request.Content,
             Items = request.Items.Select(x => new ResultItem
             {
                 FormItemId = x.FormItemId,
                 Values = x.Values.Select(v => new ResultItemValue
                 {
-                    Value = v.Value,
+                    Value = v.Value ?? string.Empty,
                 }).ToList(),
             }).ToList(),
         };
